apiForModalData
refactor: apiForModalData move to another file
